divide models
