+++ syncano/models/__init__.py
@@ -1,3 +1,8 @@
 from .base import *  # NOQA
 from .fields import *  # NOQA
+from .instances import *  # NOQA
+from .accounts import *  # NOQA
+from .billing import *  # NOQA
 from .channels import *  # NOQA
+from .classes import *  # NOQA
+from .incentives import *  # NOQA
